% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentinelSearch.R
\name{SentinelSearch}
\alias{SentinelSearch}
\title{SentinelSearch}
\usage{
SentinelSearch(musr = NA, mpass = NA, bbox = NA,
  sensing_date_start = NA, sensing_date_end = NA, platformname = NA)
}
\arguments{
\item{musr}{Copernicus Open Access Hub user name (authentication)}

\item{mpass}{Copernicus Open Access Hub password (authentication)}

\item{bbox}{Geographic area of interest (a raster extent object)}

\item{sensing_date_start}{Sensing start (from) datetime. Must be POSIXct object with tz.}

\item{sensing_date_end}{Sensing end (to) datetime. Must be POSIXct object with tz.}

\item{platformname}{Name of satellite. Either "Sentinel-1", "Sentinel-2" or c("Sentinel-1", "Sentinel-2").}
}
\value{
search response from server (json list)
}
\description{
Searches available sentinel data for an area of interest and date range
}
