% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentinelSearchToDF.R
\name{SentinelSearchToDF}
\alias{SentinelSearchToDF}
\title{SentinelSearchToDF}
\usage{
SentinelSearchToDF(resp = NA, musr = NA, mpass = NA)
}
\arguments{
\item{resp}{JSON "list" object generated from function SentinelSearch()}

\item{musr}{Copernicus Open Access Hub user name (authentication)}

\item{mpass}{Copernicus Open Access Hub password (authentication)}
}
\value{
Dataframe of search results
}
\description{
Parse Sentinel Search json list object to dataframe with key fields for each image
}
