% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SentinelThumbnailPreview.R
\name{SentinelThumbnailPreview}
\alias{SentinelThumbnailPreview}
\title{SentinelThumbnailPreview}
\usage{
SentinelThumbnailPreview(sentinel.df = NA, thumbnail.dir = NA, musr = NA,
  mpass = NA)
}
\arguments{
\item{sentinel.df}{a dataframe generated from function SentinelSearchToDF()}

\item{thumbnail.dir}{folder directory to store thumbnail images in}

\item{musr}{Copernicus Open Access Hub user name (authentication)}

\item{mpass}{Copernicus Open Access Hub password (authentication)}
}
\value{
Dataframe of search results
}
\description{
Preview small sentinel thumbnail images to determine which images are good for use and cloud free
}
