% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square A Number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
Square of the input
}
\description{
Takes in a numeric value and squares it
}
